Use a 9bit UART for receiving commands from host
Modified the Raspberry Pi RX UART PIO example to be a 9bit UART.
Recognizes reset, mouse and keyboard queries without erroring.
Have yet to test LED commands.

+++ src/firmware/lib/QuokKM5/include/log_cache.h
@@ -75,7 +75,7 @@ inline void log_raw()
 }
 
 
-extern "C" unsigned long millis();
+extern "C" uint32_t millis();
 
 // Variadic template for printing multiple items
 template<typename T, typename T2, typename... Rest>
